Revert "Polimorfismo"
This reverts commit e082aec3d3926ccb596e04113e4349186891b7a7.

+++ src/Interface/ImpactoAmbiental.java
@@ -1,5 +1,5 @@
 package Interface;
 
 public interface ImpactoAmbiental {
-    int calcularImpacto();
+    double calcularImpacto();
 }
